add app bar icon
